+++ Sistema/Usuarios/usuariosAdm.php
@@ -6,7 +6,8 @@ require '../../conexion_BD.php';
 
 require_once "../../EVENT_BITACORA.php";
 session_start();     
-$usuario=$_SESSION['usuario'];
+$usuario=$_SESSION['user'];
+$ID_Rol=$_SESSION['ID_Rol'];
 
 //Parte 2
                 
@@ -59,13 +60,18 @@ $R_F_Vencida= date("Y-m-j",strtotime($R_Fecha_actual."+ ".$diasV." days")); /*le
                   <div class="box">
                     <div class="box-header with-border">
                           <h1 class="box-title">Mantenimiento usuarios</h1>
+                          <?php $sql=$conexion->query("SELECT * FROM tbl_permisos where Permiso_Insercion=1 and ID_Rol=$ID_Rol ");
+if ($datos=$sql->fetch_object()) { ?>
                           <button class="btn btn-success" id="btnagregar" name="btnAgregar" onclick="mostrarform(true)"><i class="zmdi zmdi-account-add"></i>Agregar Usuario</button>
                           <div class="box-tools pull-right">
+                          <?php } ?>
                         </div>
                         <br>
                     </div>
                     <!-- /.box-header -->
                     <!-- centro -->
+                    <?php $sql=$conexion->query("SELECT * FROM tbl_permisos where Permiso_consultar=1 and ID_Rol=$ID_Rol ");
+if ($datos=$sql->fetch_object()) { ?>
                     <div class="panel-body table-responsive" id="listadoregistros">
                         <table id="tbllistado" class="table table-bordered table-hover">
                         
@@ -108,11 +114,17 @@ $R_F_Vencida= date("Y-m-j",strtotime($R_Fecha_actual."+ ".$diasV." days")); /*le
                               <td><?php echo $mostrar['Fecha_Vencimiento']?></td>
                               <td><?php echo $mostrar['Estado_Usuario']?></td>
                               <td>
+                              <?php $sql=$conexion->query("SELECT * FROM tbl_permisos where Permiso_Actualizacion=1 and ID_Rol=$ID_Rol");
+if ($datos=$sql->fetch_object()) { ?>
                               <a href='Update_Usuarios.php?ID_Usuario=<?php echo $mostrar['ID_Usuario']; ?>' class='boton-editar'>
                               <i class='zmdi zmdi-edit'></i> Editar
+                              <?php } ?>
                               </a>
+                              <?php $sql=$conexion->query("SELECT * FROM tbl_permisos where Permiso_Eliminacion=1 and ID_Rol=$ID_Rol ");
+if ($datos=$sql->fetch_object()) { ?>
                               <a href='Delete_Usuarios.php?Usuario=<?php echo $mostrar['Usuario']; ?>' onclick='return confirmar()' class='boton-eliminar'>
                               <i class='zmdi zmdi-delete'></i> Eliminar
+                              <?php } ?>
                               </a>
                             </td>
                              </tr>
@@ -122,6 +134,7 @@ $R_F_Vencida= date("Y-m-j",strtotime($R_Fecha_actual."+ ".$diasV." days")); /*le
                           </tfoot>
                         </table >
                     </div>
+                    <?php } ?>
                     <div class="panel-body" id="formularioregistros">
                         <form name="formulario" id="formulario" action="Insert_Usuarios.php" method="POST">
                         <div class="container">
