+++ Sistema/Voluntarios/VoluntariosAdm.php
@@ -6,7 +6,8 @@ require '../../conexion_BD.php';
 
 require_once "../../EVENT_BITACORA.php";
 session_start();     
-$usuario=$_SESSION['usuario'];
+$usuario=$_SESSION['user'];
+$ID_Rol=$_SESSION['ID_Rol'];
 
 ?>
 
@@ -45,13 +46,18 @@ $usuario=$_SESSION['usuario'];
                   <div class="box">
                     <div class="box-header with-border">
                           <h1 class="box-title">Mantenimiento Voluntarios</h1>
+                          <?php $sql=$conexion->query("SELECT * FROM tbl_permisos where Permiso_Insercion=1 and ID_Rol=$ID_Rol ");
+if ($datos=$sql->fetch_object()) { ?>
                           <button class="btn btn-success" id="btnagregar" name="btnAgregar" onclick="mostrarform(true)"><i class="zmdi zmdi-account-add"></i> Agregar Voluntario</button>
                           <div class="box-tools pull-right">
+                            <?php } ?>
                         </div>
                         <br>
                     </div>
                     <!-- /.box-header -->
                     <!-- centro -->
+                    <?php $sql=$conexion->query("SELECT * FROM tbl_permisos where Permiso_consultar=1 and ID_Rol=$ID_Rol ");
+if ($datos=$sql->fetch_object()) { ?>
                     <div class="panel-body table-responsive" id="listadoregistros">
                         <table id="tbllistado" class="table table-bordered table-hover">
                         
@@ -84,11 +90,17 @@ $usuario=$_SESSION['usuario'];
                               <td><?php echo $mostrar['Telefono_Voluntario']?></td>
                               <td><?php echo $mostrar['Direccion_Voluntario']?></td>
                               <td>
+                              <?php $sql=$conexion->query("SELECT * FROM tbl_permisos where Permiso_Actualizacion=1 and ID_Rol=$ID_Rol");
+if ($datos=$sql->fetch_object()) { ?>
                               <a href='Update_Voluntarios.php?ID_Voluntario=<?php echo $mostrar['ID_Voluntario']; ?>' class='boton-editar'>
                               <i class='zmdi zmdi-edit'></i> Editar
+                              <?php } ?>
                               </a>
+                              <?php $sql=$conexion->query("SELECT * FROM tbl_permisos where Permiso_Eliminacion=1 and ID_Rol=$ID_Rol ");
+if ($datos=$sql->fetch_object()) { ?>
                               <a href='Delete_Voluntarios.php?ID_Voluntario=<?php echo $mostrar['ID_Voluntario']; ?>' onclick='return confirmar()' class='boton-eliminar'>
                               <i class='zmdi zmdi-delete'></i> Eliminar
+                              <?php } ?>
                               </a>
                             </td>
                              </tr>
@@ -98,6 +110,7 @@ $usuario=$_SESSION['usuario'];
                           </tfoot>
                         </table>
                     </div>
+                    <?php } ?>
                     <div class="panel-body" id="formularioregistros">
                         <form name="formulario" id="formulario" action="Insert_Voluntarios.php" method="POST">
                         <div class="container">
