+++ Sistema/pagos/PagosAdm.php
@@ -13,7 +13,8 @@ require_once "../../EVENT_BITACORA.php";
                 
 $R_Fecha_actual = date('Y-m-d');       /*obtiene la fecha actual*/
 session_start();     
-$usuario=$_SESSION['usuario'];
+$usuario=$_SESSION['user'];
+$ID_Rol=$_SESSION['ID_Rol'];
 
 $sql1=$conexion->query("SELECT * FROM `tbl_ms_parametros` WHERE ID_Parametro=7");
 
@@ -61,13 +62,18 @@ $R_F_Vencida= date("Y-m-j",strtotime($R_Fecha_actual."+ ".$diasV." days")); /*le
                   <div class="box">
                     <div class="box-header with-border">
                           <h1 class="box-title">Mantenimiento de Pagos</h1>
+                          <?php $sql=$conexion->query("SELECT * FROM tbl_permisos where Permiso_Insercion=1 and ID_Rol=$ID_Rol ");
+if ($datos=$sql->fetch_object()) { ?>
                           <button class="btn btn-success" id="btnagregar" name="btnAgregar" onclick="mostrarform(true)"><i class="zmdi zmdi-account-add"></i>Agregar Pago</button>
                           <div class="box-tools pull-right">
+                            <?php } ?>
                         </div>
                         <br>
                     </div>
                     <!-- /.box-header -->
                     <!-- centro -->
+                    <?php $sql=$conexion->query("SELECT * FROM tbl_permisos where Permiso_consultar=1 and ID_Rol=$ID_Rol ");
+if ($datos=$sql->fetch_object()) { ?>
                     <div class="panel-body table-responsive" id="listadoregistros">
                         <table id="tbllistado" class="table table-bordered table-hover">
                         
@@ -108,11 +114,17 @@ $R_F_Vencida= date("Y-m-j",strtotime($R_Fecha_actual."+ ".$diasV." days")); /*le
                               <td><?php echo $mostrar['Nombre_Usuario']?></td>
                               <td><?php echo $mostrar['Fecha_de_transaccion']?></td>
                               <td>
+                              <?php $sql=$conexion->query("SELECT * FROM tbl_permisos where Permiso_Actualizacion=1 and ID_Rol=$ID_Rol");
+if ($datos=$sql->fetch_object()) { ?>
                               <a href='Update_Pago.php?ID_de_pago=<?php echo $mostrar['ID_de_pago']; ?>' class='boton-editar'>
                               <i class='zmdi zmdi-edit'></i> Editar
+                              <?php } ?>
                               </a>
+                              <?php $sql=$conexion->query("SELECT * FROM tbl_permisos where Permiso_Eliminacion=1 and ID_Rol=$ID_Rol ");
+if ($datos=$sql->fetch_object()) { ?>
                               <a href='Delete_Pago.php?ID_de_pago=<?php echo $mostrar['ID_de_pago']; ?>' onclick='return confirmar()' class='boton-eliminar'>
                               <i class='zmdi zmdi-delete'></i> Eliminar
+                              <?php } ?>
                               </a>
                             </td>
                              </tr>
@@ -122,6 +134,7 @@ $R_F_Vencida= date("Y-m-j",strtotime($R_Fecha_actual."+ ".$diasV." days")); /*le
                           </tfoot>
                         </table>
                     </div>
+                    <?php } ?>
                     <div class="panel-body" id="formularioregistros">
                         <form name="formulario" id="formulario" action="Insert_Pago.php" method="POST">
                         <div class="container">
